pyproject.toml
[project]
name = "cv-analyser"
version = "0.1.0"
description = "Analisador de currículos."
authors = [
    {name = "Messer",email = "claraborgespsic@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "google-api-python-client (>=2.179.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "tinydb (>=4.8.2,<5.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "streamlit (>=1.49.0,<2.0.0)",
    "streamlit-aggrid (>=1.1.8.post1,<2.0.0)",
    "langchain-groq (>=0.3.7,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "utilz (>=0.4,<0.5)",
    "utils (>=1.0.2,<2.0.0)",
    "pymupdf (>=1.26.4,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "chardet (>=5.2.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
